code/cloud function 2 - find words and run face detection/requirements.txt
opencv-python
google-cloud-storage
google-cloud-vision